Move Dispatchable container to Geometry where it's used
