+++ SimG4Core/Geometry/interface/DDG4DispContainer.h
@@ -1,0 +1,10 @@
+#ifndef SimG4Core_DDG4DispContainer_H
+#define SimG4Core_DDG4DispContainer_H
+
+#include "SimG4Core/Geometry/interface/DDG4Dispatchable.h"
+
+#include <vector>
+
+using DDG4DispContainer = std::vector<DDG4Dispatchable *>;
+
+#endif
