--- SimG4Core/Notification/interface/DDG4DispContainer.h
@@ -1,10 +1,0 @@
-#ifndef SimG4Core_DDG4DispContainer_H
-#define SimG4Core_DDG4DispContainer_H
-
-#include "SimG4Core/Notification/interface/DDG4Dispatchable.h"
-
-#include <vector>
-
-typedef std::vector<DDG4Dispatchable *> DDG4DispContainer;
-
-#endif
